Moved column/row OutputText() into ArxGame
The function still isn't complete, although if more features were needed it may be better to go with a full engine for gui layout. (rects, backgrounds, widgets, hierarchy, ...)

This should make it easier to find a good place to output text while dynamically resizing the canvas (screen) and font glyphs.