SDLInputBackend: Fix relative mouse input
Revert to calculating relative mouse movements from the changes in
absolute mouse positions. The SDL-provided relative movements are only
usable when grabbing the input - otherwise the (hidden) mouse will just
exit the window. SDL however doesn't provide any way to grab just the
mouse and grabbing the whole keyboard is unacceptable.
Note: SDL always grabs the keyboard in fullscreen mode, there is nothing
we can do about that.

Instead, we just re-center the mouse every frame while in mouse look
mode, as the original game used to do. While this still leaves a
possibility for the mouse to escape the window for fast movements, it
will only happen if rendering is really slow - and has never been a
problem so far.

At some point we should move that re-center logic into the individual
input backends to allow for proper (non-SDL) relative mouse input.