Re-add gamma/brightness/contrast options
Some people really need this feature, so we should work on adding it
back, maybe even before we manage to get the render ready for PP.

Reverts commit 3ad44dcd92daf60cff1c7f4d13fe20bda242ca01

This is not meant as a final version, as it has all the problems those
settings had before we removed them. These include:
 * Doesn't work under Linux at all (at least for me)
 * From the IRC logs it looks like this changed the gamma for the whole
   desktop when in windowed mode.
More tests are needed on these problems.

We may want to use those settings like they are used now, but only when
in fullscreen mode.

Either way, we should make sure that the defaults provide a 1:1 mapping.