Don't assume SDLWindow is only used with SDL input
SDLWindow can be used with both SDLInputBackend and DInput8Backend.
While DInput8Backend ignores all SDL input events, the SDLWindow has
to forward them to SDLInputBackend if that is being used.

This fixes a bug in that forwarding code that caused a crash when the
SDLWindow was paired with a DInput8Backend.

Fixes bug #290