Fix clearing IO speech on destruction
ARX_SPEECH_ClearIOSpeech() had a bug that caused the speech for
non-player IOs to not be cleared, leaving the IO speech slot in use
and possible causing illegal references to deleted IOs later on.

This could not have caused any long-lasting memory leaks as all IO
speech slots are cleared when changing levels.