Improved handling of D3D9 lost device state.
Previously, I relied on the fact that most calls to D3D should succeed even when a device is lost:
http://msdn.microsoft.com/en-us/library/windows/desktop/bb174714(v=vs.85).aspx

That still led to a few corner cases that were not handled properly... Now the game will be stuck in D3D9Window::tick() in case of a lost device, but it will keep on processing messages. This is much safer.

Fixes bug #274
Fixes bug #314