Print a message if arxcrashreporter is missing
If AL crashes and arxcrashreporter could not be found/run we send a KILL
signal to the original arx process. In most terminals this will simply
print "Killed.".

This commit adds a message informing the user of the actual signal that
caused the crash as well as asking to install arxcrashreporter.