Ignore hardcoded languages for cinematic sounds
The cinematics files (.cin) have hardcoded resource paths for English,
French, German and Spanish variants of speech sound files. (but not for
Italian and Russian)

However, only the english variant was ever used.

This also worked for the French, German and Spanish versions of the
game, as those store the localised speech files at both the correct and
'English' resource paths. Russian and Italian versions only store the
files at the correct paths.

This commit changes the .cin file loading code to completely ignore any
hardcoded languages in the sound paths and instead always use the one
for the current language.

The original game had similar code, but it was apparnetly removed at
some code, which is not surprising considering how messy it was.

Fixes speech during cinematics for the Italian and Russian version.

Fixes bug #335