Always use user/config location in home directory
Previously, if no data and user directories were found, the user dir was
set to the CWD. Similarily, the CWD was used as the config diretcory if
no data, user and config directories were found.

This commit changes it so that the config and user directories are
always created at the proper location in the user's home directory.

They can still be set to the current directory using the --user-dir and
--config-dir command-line options.