install-verify: support a bundled version of AF
The only difference is that the speech.pak differs from the normal
german version: all files in 'speech/deutsch' match, but the files
in 'speech/francais' and 'speech/english' differ (both contents and
which files are present). In the normal german version these directories
are just copies of 'speech/deutsch'.

We'll have to see if this speech.pak is also used for non-german
releases (or maybe the bundled version was multilang?) and thus would
need a different loc.pak