Remove redundant parameter from TREATZONE_AddIO()
The function was always passed an Entity pointer as well as a redundant
entity index. This was presumably done because entity indices were
expensive to calculate in the original AF code (O(total_entities)).

AL caches this index in the Entity object, so there is no performance
impact in just passing the pointer and getting the index from that.