Fix build with CMake 2.8.10
Previously we checked for SDL by using
  find_package(SDL 1.2 EXACT)

This worked with earlier CMake versions as the find script for SDL
didn't provide any version, so CMake just ignored the '1.2 EXACT' part.

In CMake 2.8.10, FindSDL.cmake reports the SDL version number as
major.minor.patch, which is never exactly equal to 1.2.

This commit relaxes the check by removing the 'EXACT' requirement.