Fix rendering of umlauts and accented characters
In commit 6d93bc74 I decreased the upper limit for which characters
to load at font object creation but forgot to adjust the respective
lower limit for which characters we try to load on demand.

This caused all characters in the [U+7F, U+FF] range to not be rendered.