Fix arx-install-data with mksh
Consider the following code:

set -- foobar
for item in ; do echo "$item" ; done

This prints foobar in mksh but nothing in bash, dash, zsh and ksh.
My reading of the POSIX shell spec tells me that mksh is wrong, but
it's not hard to change the code so it also works on mksh by preventing
those weird for loops.