RAF cheat: Also increase the player caster level
The previous commit fixes the RAF cheat so that it doesn't limit the
player caster level to 1.

It now applies the caster level bonus of the cheat to when casting
previously prepared spells or when NPCs cast spells, however not when
the player casts spells directly.
This commit fixes that inconsistency.