Changes to aura controls.
Makes the aura control method for the "Armor" spell analogous to the one used for the "Protection from fire" spell.

Turns on the aura for the "Lower armor" spell if the spell is active but the aura is disabled. If a protection spell and "Lower armor" are active at the same time on the same target, the lower armor aura is hidden. Previously, the lower armor aura did not appear when protection ended before the end of "Lower armor"; this is now fixed.