SDL*Window: Verify that we get an accelerated ctxt
This is supposed to weed out "bad" graphics drivers that SDL somehow
picks up if we request an MSAA level higher than what the better driver
supports - namely the "GDI Generic" driver under Windows.

Also verify that we actually get a backbuffer with the requested number
of MSAA samples for the same reason.