Fix player hands and equipment clipping with walls
Arx Fatalis used a trick to draw the player and quipped items in front
of everything else even when they actually intersect. While not
realistic, this is a common technique in first-person games as it avoids
clipping artifacts when the player gets close to walls (while not
excessively restricting movement by increasing the player's collision
shape).
This was broken in commit 6e4e2f (for the player body) and commit 256d14
(for equipped items).

This commit re-implements the trick by using a by first rendering the
player with glDepthFunc(GL_ALWAYS) before the normal player rendering.
This differs a bit from the original solution which pushed the vertices
to the font as they allowed *some* intersection hand had different push
factors for the player body, feet and equipped items.

A proper solution would be to adjust the weapon and hand positions to
not intersect with the rest of the scene, but that is much more complex.