Set SHOW_FLAG_DESTROYED in Entity::destroy()
Previously it set SHOW_FLAG_KILLED if the entity wasn't created by a
script and *deleted* otherwise. The only reason entities not loaded by
scripts aren't immediately destroyed is that they need to be written to
the save file so that we can remember their destruction. This is needed
because they are still referenced by the level file and because we don't
want to reuse their instance number as that may be associated with a
special script that should not be used for random objects of the class.

The KILLED flag will however not prevent the entities from receiving
script events and will cause them to be needlessly loaded again from
save file, bloating the runtime entity list.