Don't freeze broken scripts
Instead, continue with the next line.

Our script parsing is slightly stricter than the original AF and doesn't
allow commands to span multiple lines. This only affects a few scripts
with newlines between the wrong tokens, some of which already have
script warning suppressions - and we already continued with the next
line if there was a script warning suppression defined.

These peculiar newline positions should be fixed in the actual scripts
once we have the infrastructure for that. Until then, they can be
collected as script warning suppressions.

Fixes: bug #584