Fix entity movement not being set if ftr is zero
Original AF updated the movement vector only if there was a valid
animation. This was changed in commit a909fe to only update if there was
a valid animation *and* the calculated value was not the zero vector.

This commit changes it to *always* update the movement vector.

Fixes a spot in the intro sequence where a door was directly in front of
the camera.