Fix crash when summon/raise dead spell ends early
In commit e04954 / commit a68c1b the position of the end sound was
changed to the current position of the summoned creature for the
summon / raise dead spells respectively. If no creature was summoned yet
(or in case of the summon spell, if the summoned creature was't a demon)
this could lead to an invalid entity reference that wasn't checked for.

This commit changes it to not play the (unsummon?) sound if no creature
was summoned.

Fixes: bug #589