Change of the replacing a player's broken weapon by another from the player's inventory.
Previously, weapons were replaced by equally powerful ones if available, otherwise by the weakest of the weaker ones or by the strongest of the stronger ones, in that order. Now, the priorities are different: equally powerful, then the strongest of the weaker ones, then the weakest of the stronger ones.