Use alpha to coverage to AA color-keyed textures
This currently only works if alpha blending isn't enabled.
When we are already alpha blending we should just include the src alpha
from the color-keyed texture in the blend factor.

Performance impact should be minor, but we may still want to add a
separate (from base AA) config option for alpha to coverage.

Also, we currently (almost) always enable the ColorKey render state,
even if no color keyed texture is being used. This is a bit wasteful.