Never add -fPIE to global CXX flags
Qt5 needs the arxcrashreporter code to be compile with -fPIE.
libArxIO needs to be build with -fPIC

GCC <= 4.7 don't like it when you use both of these flags at the
same time.

Change it so that -fPIE is only used on sources unique to the crash
reporter - ie the sources using Qt functions. This seems to be enough.