Get data directory from "data" file/dir in exe dir
If "data" is a file, each line specifies a data directory to add.
Non-absolute paths are resolved relative to the exe directory.

This has two uses:
- The data file is used to tell the arx binary in the build dir where
  to load the data files from the source tree without installing them.
- In (portable) installs the data files can be positioned relative
  to the arx binary without needing to be in the same directory

Paths found in this way are added directly before the exe dir itself.

This is all done for both ${arx_PATH} and the actual exe dir.