Create the Window as soon as possible
This makes the regular startup as responsive as possible.

This will mean that, if there are missing data files, the (fullscreen)
window flashes briefly before the error message is shown. With the
default config (fullscreen windowed) this should not not involve any
mode changes.

I think we should optimize for the non-error case.