CMake: Fix C++ version check
Features of higher C++ versions than available were tested.

This also adds support for the CXX_STANDARD property introduced in
CMake 3.1. This is useful as a fallback for unknown compilers but is
also required to prevent CMake from requesting a lower version if we
link a library that requests a C++11 compile feature.