Fix wrong path waypoint event
When a frame hit the target time of a path waypoint exactly we
incorrectly reported the next waypoint index as reached.

Additonally, when loading old saves this could mean that we
send the 'on pathend' script event again for paths that were already
completed: Even though the path time was past the end before saving,
the millisecond resolution in the save file could result in it
being rounded down to match the end exactly after loading.

Fixes: bug #1293