Trail: Reimplement arrow/speed trail effect
This makes the appearance independent of the view direction and smoothly
fades out the trail towards the front and edges while fixing the size
and color interpolation when the trail does not have the full number
of segments.

It also no longer adds new positions when the trail is not moving and
instead removes old positions, allowing users to check when the trail
has fully disappeared.