Core: Implement an improved bow aim
Previously arrows were always show in the direction the player is
looking at starting from around the player center.

The new bow aim shoots the array from the actual position on the bow
in the actual direction of the arrow on the bow and instead aims the bow
at whatever the player is looking at.