Text: Use arx_base.ttf
This is new in AF 1.22 and is the same as arx_russian.ttf included
in 1.21.

arx_base.ttf has cyrillic glypths in positions where they should not
be which breaks things for French so only use it as a fallback.

For the Russian version, arx_russian.ttf is preferred anyway.