Text: Fix line length calculation
The numChars return from ARX_UNICODE_FormattingInRect() did not include
the last character when the whole text fit into rect but no more lines
would fit.

The credits code had a workaround for this which did not take into
account that characters can be longer than one byte.

This fixes the bug in ARX_UNICODE_FormattingInRect() and removes the
credits workaround.

Fixes: issue #1545