Dragging: Unstack items when thrown or dropped on the floor
In commit ec7103b, dropping stacks of items was unintentionally made
possible. While dropping stacks is desirable, this causes problems
with scripts that do not expect there to be stacks outside containers.
Specifically, some raw foods turn of interactivity while cooking
but then only replace one item in the stack when cooking is done,
leaving the rest of the stack uncooked and non-interactive.
Fix this by splitting the stack into individual items when dropped.

Fixes: issue #1522
Implements: issue #36