WAV: Limit the reported file size to the available data
We use the sample count from the 'fact' chunk to determine the playback
length for ADPCM-aligned files because ADPCM can only encode a multiple
of the block size and thus may have silent padding at the end.
However, for some files the sample count is higher than the number of
samples encoded in the 'data' chunk and we ended up trying to read past
the end of the 'data' chunk when decoding.