cmake: Properly save/restore CMAKE_FIND_LIBRARY_SUFFIXES
CMAKE_FIND_LIBRARY_SUFFIXES has different behavior when undefined and
when defined but empty:
Empty means to use an empty suffix while undefined means to use a
hardcoded default for the platform we are running on.

Unfortunately, set(a ${b}) will undefine a when b is empty. On the other
hand, set(a "${b}") will always define a to be empty but defined so we
have to do this dance to fully preserve the state of this variable.