Platform: Don't use Clang's __builtin_assume()
Clang will discard the entire expression (with a warning) if it calls
any functions, including custom operators, that have not been manually
marked with the [[gnu::pure]] or [[gnu::const]] attributes.

Since we already use a conditional __builtin_unreachable() with GCC for
this (which unfortunately does not guarantee that no code is emitted)
we can just use that for Clang as well.