ChangeLevel: Fix out of bounds access to NPC extra_rotate data in saves
This affects both reading and writing saves so could corrupt both the
rotation angles as well as the first stacked target. The write also
touches the saved blood color but that is overwritten with the correct
value later.

Only affects NPCs with extra rotation data - i.e. those that hat the
look around (-l) or stare at (-a) behaviors active at some point.

Was broken in commit 239cff7.

Fixes: issue #1668